added detailvc tableview
moved assignmentLong to tableview under info header
tableview now shows assignmentLong + downloads + links under different headers
according to what each is. seemed like the easiest way to do this given that
the tableview implements scrolling for us and getting scrolling to work was
going to be a pain otherwise (at least i couldn't do it). perhaps will revise
in the future to not show headers of empty sections